translator - further separate some pieces based on the work done
